+++ src/routes.js
@@ -73,8 +73,9 @@ export default function getRoutes(
 		method: ['GET', 'POST', 'DELETE', 'PATCH'],
 		path: '/api',
 		handler: (request, reply) => {
-			const queryResponses$ = request.authorize()  // ensures that a valid oauth_token is present in the request
+			const queryResponses$ = request.authorize()  // ensure a valid oauth_token is present in the request
 				.flatMap(proxyApiRequest$);
+
 			queryResponses$.subscribe(
 				queryResponses => {
 					reply(JSON.stringify(queryResponses)).type('application/json');
